Move more files into diff/
