+++ src/diff/mod.rs
@@ -3,3 +3,4 @@ mod graph;
 pub mod myers_diff;
 pub mod sliders;
 pub mod unchanged;
+pub mod changes;
